==> Data_Aggregation/5_analysis_data_aggregation.ipynb <==
import os
import pandas as pd

analysis_folder = r"C:\Users\yegor\Desktop\Honors Thesis\Data\Analysis\Analysis_Data"
os.chdir(analysis_folder)

mo = pd.read_csv("monthly_overall.csv",    dtype={
        "fips_standard": str,
        'Urban': str,
        "year": int,
        "month": int,
        "total_deaths": int},low_memory=False)     # fips_standard, Urban, year, month, total_deaths
mo['fips_standard']= mo['fips_standard'].astype(str).str.zfill(5)


ma = pd.read_csv("monthly_age.csv", dtype={
        "fips_standard": str,
        'Urban': str,
        "year": int,
        "month": int,
        'age_bin': str,
        "deaths_age": int},low_memory=False)         # fips_standard, Urban, year, month, age_bin, deaths_age
ma['fips_standard']=ma['fips_standard'].astype(str).str.zfill(5)

ms = pd.read_csv("monthly_sex.csv", dtype={
        "fips_standard": str,
        'Urban': str,
        "year": int,
        "month": int,
        "sex":str,
        "deaths_gender": int},low_memory=False)         # ... sex, deaths_gender
ms['fips_standard']= ms['fips_standard'].astype(str).str.zfill(5)

mr = pd.read_csv("monthly_race.csv", dtype={
        "fips_standard": str,
        'Urban': str,
        "year": int,
        "month": int,
        'race': str,
        "deaths_race": int},low_memory=False)        # ... race, deaths_race
mr['fips_standard']= mr['fips_standard'].astype(str).str.zfill(5)
me = pd.read_csv("monthly_education.csv", dtype={
        "fips_standard": str,
        'Urban': str,
        "year": int,
        "month": int,
        'education': str,
        "deaths_education": int},low_memory=False)    # ... education, deaths_education
me['fips_standard']=me['fips_standard'].astype(str).str.zfill(5)





mr

# 2) Read the population long file and pull out only the TOTAL rows
pop = pd.read_csv(
    "population_counts.csv",
    dtype={
        "fips": str,
        "year": int,
        "Attribute": str,
        "Population": int
    },
    low_memory=False
)

pop

# overall
mo2 = mo.copy()
mo2["Attribute"] = "Total"
mo2 = mo2.rename(columns={"total_deaths":"Deaths"})
mo2 = mo2[["fips_standard","Urban","year","month","Attribute","Deaths"]]

# age
ma2 = ma.copy()
ma2 = ma2.rename(columns={"age_bin":"Attribute","deaths_age":"Deaths"})
ma2 = ma2[["fips_standard","Urban","year","month","Attribute","Deaths"]]

# sex
ms2 = ms.copy()
ms2 = ms2.rename(columns={"sex":"Attribute","deaths_gender":"Deaths"})
ms2 = ms2[["fips_standard","Urban","year","month","Attribute","Deaths"]]

# race
mr2 = mr.copy()
mr2 = mr2.rename(columns={"race":"Attribute","deaths_race":"Deaths"})
mr2 = mr2[["fips_standard","Urban","year","month","Attribute","Deaths"]]

# education
me2 = me.copy()
me2 = me2.rename(columns={"education":"Attribute","deaths_education":"Deaths"})
me2 = me2[["fips_standard","Urban","year","month","Attribute","Deaths"]]


mort_long = pd.concat([mo2, ma2, ms2, mr2, me2], ignore_index=True)


mort_long.rename(columns={"fips_standard": "fips"}, inplace=True)


mort_long

pop

# pop has fips_standard, year, Attribute, Population
df = mort_long.merge(
    pop,
    on=["fips","year","Attribute"],
    how="left"
)

# compute rate per million
df["rate_per_million"] = df["Deaths"] / df["Population"] * 1_000_000


df = df.sort_values(
    by=["fips", "year", "month",'Attribute'],
    ascending=[True, True, True, True]
).reset_index(drop=True)


df

df_clean = df.dropna()
df_clean

df_temp = pd.read_csv(
    "county_summer_EHDs_1989_2019.csv",
    usecols=["GEOID", "month", "year", "is_EHD"],   # drop NAME
    dtype={
        "GEOID": str,
        "month": int,
        "year": int,
        "is_EHD": int
    },
    low_memory=False
)

# rename and zero‑pad to 5 digits
df_temp = df_temp.rename(columns={"GEOID": "fips"})
df_temp = df_temp.rename(columns={"is_EHD": "EHDs"})

df_temp["fips"] = df_temp["fips"].str.zfill(5)

df_temp

df_merged = df_clean.merge(
    df_temp[["fips", "year", "month", "EHDs"]],
    on=["fips", "year", "month"],
    how="left"
)




df_merged

df_merged["post_icd10"]=(df_merged['year']>=1999).astype(int)

df_merged

df_clean_2 = df_merged.dropna()
df_clean_2.to_csv(r"analysis_data.csv", index=False)


df_clean_2

total_data = df_clean_2[df_clean_2['Attribute'] == 'Total']
total_data= total_data[total_data['Urban']== '0.0']

# Step 3: Sum all deaths and all population exposures
deaths_sum = total_data['Deaths'].sum()
population_sum = total_data['Population'].sum()

# Step 4: Calculate baseline rate per million
baseline_rate_per_million_per_month = (deaths_sum / population_sum) * 1_000_000

# Step 5: Print the results
print(f"Total deaths: {deaths_sum}")
print(f"Total population exposure: {population_sum}")
print(f"Baseline suicide rate per million per month: {baseline_rate_per_million_per_month:.2f}")
